refactor(roll): lift interfaces into the root mod file
