+++ src/roll/resistance.rs
@@ -1,0 +1,5 @@
+use super::ResistanceOutcome;
+
+trait Resistance {
+    fn roll(&self, n: u8) -> ResistanceOutcome;
+}
